fix(my-actions,-profile/page.tsx,-summary.tsx): renamed my-actions folder to profile, updated function to return data summary for user
